+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---------------
 - add line number and word-wrap toggles to more controls
 - add diagram export to Tools-menu
+- add option to export diagram with selected entities
 - add YAML formatting options to value editor
 - retain undo history when loading SQL from file to SQL editor
 - fix error on changing column order in trigger/view editor